+++ CHANGELOG.md
@@ -11,6 +11,14 @@
 [`color.whiteness()`]: https://sass-lang.com/documentation/modules/color#whiteness
 [`color.blackness()`]: https://sass-lang.com/documentation/modules/color#blackness
 
+* Add `$whiteness` and `$blackness` parameters to [`color.adjust()`],
+  [`color.change()`], and [`color.scale()`] to modify a color's [HWB] whiteness
+  and blackness components.
+
+[`color.adjust()`]: https://sass-lang.com/documentation/modules/color#adjust
+[`color.change()`]: https://sass-lang.com/documentation/modules/color#change
+[`color.scale()`]: https://sass-lang.com/documentation/modules/color#scale
+
 ### Dart API
 
 * Add [HWB] support to the `SassColor` class, including a `SassColor.hwb()`